Extract server-specific code from runner.rs into src/server/ submodules
Move ~1100 lines of server-specific code from runner.rs into four new
modules under src/server/: data_source.rs (ServerDataSource), session.rs
(HegelSession, ServerTestRunner, parse_version), process.rs (process
management, log handling, error messages), and runner.rs (panic hooks,
test case execution, server_run entrypoint). Strip all #[cfg(not(feature
= "native"))] annotations from the extracted code since the entire
server module is already behind a cfg gate. Split embedded tests into
per-module test files under tests/embedded/server/.
